feat: SingleImageUpload image in webiny-admin (with FM)
