+++ packages/webiny-admin/src/components/index.js
@@ -9,4 +9,4 @@ export type { WithSnackbarProps } from "./withSnackbar";
 export { withDialog } from "./withDialog";
 export type { WithDialogProps } from "./withDialog";
 
-export { FileManager } from "./FileManager";
+export { default as FileManager } from "./FileManager";
