--- packages/webiny-app/src/components/SingleImageUpload.js
@@ -1,17 +1,0 @@
-// @flow
-import * as React from "react";
-import { Image } from "./Image";
-import { SingleImageUpload } from "webiny-ui/ImageUpload";
-
-type Props = Object;
-
-export default ({ imagePreviewProps, ...singleImageUploadProps }: Props) => {
-    return (
-        <SingleImageUpload
-            renderImagePreview={(renderImageProps: Object) => (
-                <Image {...renderImageProps} {...imagePreviewProps}/>
-            )}
-            {...singleImageUploadProps}
-        />
-    );
-};
