+++ packages/webiny-app/src/components/index.js
@@ -7,6 +7,4 @@ export { withKeyHandler } from "./withKeyHandler";
 export { withUi } from "./withUi";
 export type { WithDataListProps, SearchParams, WithDataListParams } from "./withDataList";
 export { Image } from "./Image";
-export { default as SingleImageUpload } from "./SingleImageUpload";
-export { default as MultiImageUpload } from "./MultiImageUpload";
 export { default as Addons } from "./Addons";
